fix: reloading config after k8s configmap replacement
Viper's WatchConfig() function already includes logic to determine if
the config file was changed after the fsnotify event or not. The added
filtering for only fsnotify.Write operations was breaking config
reloading in cases where the config file is symlinked to elsewhere, like
when using kubernetes ConfigMaps or a git-sync sidecar container.

The reason why looking for write events causes us to miss symlink
changes is due to how viper detects this case. Viper watches the
directory containing the config file and calls filepath.EvalSymlinks to
get the true path of the config file after each and every fsnotify event
it receives. If the true path has changed since the last event, it
updates the true path and calls our listener. This means viper sometimes
detects the symlink change while processing some event ahead of the
actual modification of the link. In that case, while the callback
invocation does represent a real change to the config file, the event
passed with it may be some seemingly unrelated operation (like
kubernetes writing or chmod-ing the updated target data in the config
map mount directory prior to updating the link).