Addition of some golang high level unit tests for the backend. #404
There are wider changes here then I would have ideally liked to have made, but I needed to be able to create the webservers and shut them down again to test the applications backend. I'm happy to remove these changes as I'm not really sure the best way to add tests for this code otherwise, as there is a lot of cross-dependencies between the /api/, /websocket and webui for working on a subpath.